# api/taskService.swagger.yml
swagger: '2.0'
info:
  version: 1.0.0
  title: Fetchtask handling service
produces:
  - application/json
consumes:
  - application/json

paths:
  /sendTask:
    post:
      summary: create fetchTask
      operationId: createFetchTask
      produces:
        - application/json
      parameters:
        - in: body
          name: "task"
          description: "parameters for task"
          required: true
          schema:
            type: object
            properties:
              method:
                type: string
              path:
                type: string
              headers:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
              Body:
                type: string
      responses:
        200:
          description: scheme in json
          schema:
            $ref: "#/definitions/FetchTask"
        '404':
          description: Requested page not found
        '500':
          description: Techical probles with server.
  /getTasks:
    get:
      summary: return tasks
      operationId: getAllTasks
      produces:
        - application/json
      responses:
        200:
          description: scheme in json
          schema:
            type: array
            items: {
              $ref: "#/definitions/FetchTask"
            }
        '500':
          description: Techical problems with server.
  /task/{taskID}:
    delete:
      summary: delete task
      operationId: deleteFetchTask
      produces:
        - application/json
      parameters:
        - name: taskID
          in: path
          description: ID of task to delete
          required: true
          type: integer
          format: int
      responses:
        '200':
          description: Operation succesful
        '404':
          description: A task with the current ID was not found.
        '500':
          description: No connection to database.
    get:
      summary: get task
      operationId: getTask
      produces:
        - application/json
      parameters:
        - name: taskID
          in: path
          description: ID of task to get
          required: true
          type: integer
          format: int
      responses:
        200:
          description: scheme in json
          schema:
           $ref: "#/definitions/FullTask"
        '404':
          description: A task with the current ID was not found.
        '500':
          description: Techical problems with server.

definitions:
  TaskResponse:
    type: object
    properties:
      id:
        type: integer
        format: int
      httpStatus:
        type: integer
        format: int
      method:
        type: string
      path:
        type: string
      headers:
        type: object
        additionalProperties:
          type: array
          items:
            type: string
      bodyLenght:
        type: integer
        format: int
  FetchTask:
    type: object
    properties:
      id:
        type: integer
        format: int
      method:
        type: string
      path:
        type: string
      progress:
        type: string
      body:
        type: string
      headers:
        type: object
        additionalProperties:
          type: array
          items:
            type: string
  FullTask:
    type: object
    properties:
      request:
        type: object
        properties:
          id:
            type: integer
            format: int
          method:
            type: string
          path:
            type: string
          progress:
            type: string
          body:
            type: string
          headers:
            type: object
            additionalProperties:
              type: array
              items:
                type: string
      response:
        type: object
        properties:
          id:
            type: integer
            format: int
          httpStatus:
            type: integer
            format: int
          method:
            type: string
          path:
            type: string
          headers:
            type: object
            additionalProperties:
              type: array
              items:
                type: string
          bodyLenght:
            type: integer
            format: int