# internal/models/structures.go
package models

import models2 "httpService/service/models"

const (
	StatusNew        = "New"
	StatusInProgress = "InProgress"
	StatusCompleted  = "Completed"
)

type FetchTask struct {
	ID      int                 `json:"task_id"`
	Method  string              `json:"method"`
	Path    string              `json:"path"`
	Headers map[string][]string `json:"headers"`
	Body    string              `json:"body"`
	Status  string              `json:"status"`
}

func (task *FetchTask) ConvertToSwaggerModel() *models2.FetchTask {
	return &models2.FetchTask{
		Method:   task.Method,
		ID:       int64(task.ID),
		Body:     task.Body,
		Progress: task.Status,
		Headers:  task.Headers,
		Path:     task.Path}
}

type TaskResponse struct {
	ID          int                 `json:"resp_id"`
	Status      int                 `json:"status"`
	Method      string              `json:"method"`
	Path        string              `json:"path"`
	Headers     map[string][]string `json:"headers"`
	BodyLen     int                 `json:"body_len"`
	FetchTaskID int                 `json:"ft_id"`
	Err         string              `json:"err"`
}

func (tr *TaskResponse) ConvertToSwaggerModel() *models2.TaskResponse {
	return &models2.TaskResponse{
		ID:         int64(tr.FetchTaskID),
		BodyLenght: int64(tr.BodyLen),
		HTTPStatus: int64(tr.Status),
		Path:       tr.Path,
		Method:     tr.Method,
		Headers:    tr.Headers,
	}
}